move feature assertions to steps.py
